% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulatory_monitors_co.R
\name{.define_co_column_patterns}
\alias{.define_co_column_patterns}
\title{Define column patterns to use for selecting columns}
\usage{
.define_co_column_patterns(year, level = c("site", "monitor"))
}
\arguments{
\item{year}{The year of the data}

\item{level}{The level of data (site or monitor)}
}
\value{
A character vector of column patterns
}
\description{
This function defines column patterns to use for selecting columns from the
regulatory monitor data. The column patterns are defined based on the year
and the level of data (site or monitor).
}
\keyword{internal}
