% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_monitors_data.R
\name{get_all_monitors}
\alias{get_all_monitors}
\title{Get data for all AQS monitors for a given range of years and parameter codes}
\usage{
get_all_monitors(start, end, parameter_codes)
}
\arguments{
\item{start}{The first year of the range}

\item{end}{The last year of the range}

\item{parameter_codes}{A vector of AQS parameter codes}
}
\value{
A tibble with columns for state code, county code, site number, POC, parameter code, parameter name, latitude, longitude, and year
}
\description{
This function downloads and imports data for all AQS monitors for a given range of years and parameter codes.
}
